Fix supernet-calculations.ts
Invert the loop to correctly find the smallest CIDR.

Before this change, the loop in findSupernet starts at cidr = 1 and stops at the first CIDR that encompasses all addresses. Since /1 is large enough to cover any IPv4 range, it satisfies the condition (supernetStart ≤ minAddress and supernetEnd ≥ maxAddress) and is incorrectly selected.